requirements.txt
praw>=7.6.0
python-dotenv>=0.19.0
pyttsx3>=2.90
moviepy>=1.0.3
numpy>=1.21.0
Pillow>=8.3.0
google-api-python-client>=2.34.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.6
google-auth>=2.3.0
